+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - `--diff` now also works against class directories, jars, URI:s and maven coordinates. This
   means that it is no longer necessary to `--save` the system model before running `--diff`, e.g.:
   `sift --template sift -f net.onedaybead.sift:core:0.9.0 --diff net.onedaybead.sift:core:0.7.0`
+- `--maven-repository` option has been added to specify additional Maven repositories for downloading artifacts.
+  By default, Maven Central and local user repositories are always included.
 - Remaining usages of `org.objectweb.asm.Type` replaced with `sift.core.dsl.Type`.
 - DSL: The `fork` actions displayed when using the `--profile` flag now incorporate the
   labels from the `scope()` functions.